change get size

+++ Asteroid UET/entity.h
@@ -12,7 +12,7 @@ public:
 	SDL_Texture* getTex();
 	SDL_Rect getCurrentFrame();
 	//void setRect(SDL_Rect& cn);
-	float height, width;
+	int height, width;
 	void getSize(SDL_Texture* t);
 protected:
